check external in user model
