+++ etf/evaluation/restrict_email.py
@@ -1,17 +1,6 @@
 from django.conf import settings
 from django.forms import ValidationError
 
-from etf.allowed_domains import CIVIL_SERVICE_DOMAINS
-
-
-def is_civil_service_email(email):
-    allowed = False
-    email = email.lower()
-    email_split = email.split("@")
-    if email_split[-1] in CIVIL_SERVICE_DOMAINS:
-        allowed = True
-    return allowed
-
 
 def clean_email(email):
     email = email.lower()
