+++ etf/evaluation/schemas.py
@@ -1,6 +1,6 @@
 from marshmallow import Schema, ValidationError, fields, validate
 
-from etf.evaluation.restrict_email import is_civil_service_email
+from etf.evaluation.utils import is_civil_service_email
 
 from . import choices
