+++ etf/evaluation/utils.py
@@ -7,6 +7,8 @@ import marshmallow
 from django.http import Http404
 
 from . import choices, models
+from etf.allowed_domains import CIVIL_SERVICE_DOMAINS
+
 
 event_names = set()
 
@@ -249,3 +251,12 @@ def check_evaluation_view_permission(func):
             return func(request, *args, **kwargs)
 
     return wrapper
+
+
+def is_civil_service_email(email):
+    allowed = False
+    email = email.lower()
+    email_split = email.split("@")
+    if email_split[-1] in CIVIL_SERVICE_DOMAINS:
+        allowed = True
+    return allowed
